获取临时会议Id出现错误，验证非空。

+++ src/main/java/com/ykh/conference/service/impl/ConferenceSeedServiceImpl.java
@@ -15,9 +15,9 @@ public class ConferenceSeedServiceImpl implements ConferenceSeedService {
     ConferenceSeedDao conferenceSeedDao;
     @Override
     public synchronized Integer getConfTempId(Integer confId, String name) {
-        ConferenceSeed conferenceSeed=new ConferenceSeed(confId,name);
+       ConferenceSeed conferenceSeed=new ConferenceSeed(confId,name);
        Integer max =conferenceSeedDao.findMaxId();
-        if (max >=Integer.MAX_VALUE-1){
+        if (max!=null&&max >=Integer.MAX_VALUE-1){
             conferenceSeedDao.excuteNative("truncate Table conference_seed");
         }
         conferenceSeedDao.save(conferenceSeed);
